requirements.txt
fastapi>=0.115,<1.0
uvicorn>=0.30,<1.0
pydantic>=2.8,<3.0
openai>=1.40,<2.0
python-dotenv>=1.0,<2.0
